requirements.txt
# Web framework
streamlit~=1.39.0

# Image processing
opencv-python-headless~=4.10.0
Pillow~=10.4.0
pytesseract

# Machine learning and data processing
numpy~=1.26.4
roboflow~=1.1.47
inference-gpu

# Data visualization
matplotlib~=3.9.2

# Utility libraries
attrs~=24.2.0
Jinja2~=3.1.4
python-dotenv~=1.0.1
packaging~=24.1
pyparsing~=3.1.4
typing_extensions~=4.12.2

# Web server
tornado~=6.4.1

# Serialization
protobuf~=5.28.2
